feat: Pydantic AI support
Add support for the Pydantic AI integration.

Add missing background change confirmation.

Increase agentic generation ui page size to allow it to display results
from the Pydantic AI integration better.

Add a description for the human in the loop generate_task_steps tool, so
the Pydantic AI example works as expected.

Merge write_document and confirm_changes tool into one, which allows it
work correctly with the Pydantic AI example, without the need for the
agent to perform two separate tool calls.

Sort the features, so they are easier to find when reading the code and
implementing new integrations.

Fixes: #5