feat: PydanticAI support
Add support for the PydanticAI integration.

Add missing background change confirmation.

Increase agentic generation ui page size to allow it to display results
from the PydanticAI integration better.

Add a description for the human in the loop generate_task_steps tool, so
the PydanticAI example works as expected.

Merge write_document and confirm_changes tool into one, which allows it
work correctly with the PydanticAI example, without the need for the
agent to perform two separate tool calls.

Sort the features, so they are easier to find when reading the code and
implementing new integrations.

Fixes: #5